pkg/serverless: fix dependency on trace-agent config
