+++ cmd/trace-agent/config/config_test.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package main
+package config
 
 import (
 	"errors"
@@ -217,7 +217,7 @@ func TestConfigHostname(t *testing.T) {
 		defer func() {
 			fallbackHostnameFunc = os.Hostname
 		}()
-		_, err := loadConfigFile("./testdata/site_override.yaml")
+		_, err := LoadConfigFile("./testdata/site_override.yaml")
 		assert.Contains(err.Error(), "nor from OS")
 	})
 
@@ -229,7 +229,7 @@ func TestConfigHostname(t *testing.T) {
 			t.Skip()
 		}
 		assert := assert.New(t)
-		cfg, err := loadConfigFile("./testdata/site_override.yaml")
+		cfg, err := LoadConfigFile("./testdata/site_override.yaml")
 		assert.NoError(err)
 		assert.Equal(host, cfg.Hostname)
 	})
@@ -237,7 +237,7 @@ func TestConfigHostname(t *testing.T) {
 	t.Run("file", func(t *testing.T) {
 		defer cleanConfig()()
 		assert := assert.New(t)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("mymachine", cfg.Hostname)
 	})
@@ -249,7 +249,7 @@ func TestConfigHostname(t *testing.T) {
 		err := os.Setenv("DD_HOSTNAME", "onlyenv")
 		defer os.Unsetenv("DD_HOSTNAME")
 		assert.NoError(err)
-		cfg, err := loadConfigFile("./testdata/site_override.yaml")
+		cfg, err := LoadConfigFile("./testdata/site_override.yaml")
 		assert.NoError(err)
 		assert.Equal("onlyenv", cfg.Hostname)
 	})
@@ -261,7 +261,7 @@ func TestConfigHostname(t *testing.T) {
 		err := os.Setenv("DD_HOSTNAME", "envoverride")
 		defer os.Unsetenv("DD_HOSTNAME")
 		assert.NoError(err)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("envoverride", cfg.Hostname)
 	})
@@ -356,7 +356,7 @@ func TestSite(t *testing.T) {
 	} {
 		t.Run(name, func(t *testing.T) {
 			defer cleanConfig()()
-			cfg, err := loadConfigFile(tt.file)
+			cfg, err := LoadConfigFile(tt.file)
 			assert.NoError(t, err)
 			assert.Equal(t, tt.url, cfg.Endpoints[0].Host)
 		})
@@ -552,7 +552,7 @@ func TestNormalizeEnvFromDDEnv(t *testing.T) {
 			err := os.Setenv("DD_ENV", in)
 			defer os.Unsetenv("DD_ENV")
 			assert.NoError(err)
-			cfg, err := loadConfigFile("./testdata/no_apm_config.yaml")
+			cfg, err := LoadConfigFile("./testdata/no_apm_config.yaml")
 			assert.NoError(err)
 			assert.Equal(out, cfg.DefaultEnv)
 		})
@@ -573,7 +573,7 @@ func TestNormalizeEnvFromDDTags(t *testing.T) {
 			err := os.Setenv("DD_TAGS", in)
 			defer os.Unsetenv("DD_TAGS")
 			assert.NoError(err)
-			cfg, err := loadConfigFile("./testdata/no_apm_config.yaml")
+			cfg, err := LoadConfigFile("./testdata/no_apm_config.yaml")
 			assert.NoError(err)
 			assert.Equal(out, cfg.DefaultEnv)
 		})
@@ -593,7 +593,7 @@ func TestNormalizeEnvFromConfig(t *testing.T) {
 	} {
 		t.Run("", func(t *testing.T) {
 			defer cleanConfig()()
-			cfg, err := loadConfigFile(cfgFile)
+			cfg, err := LoadConfigFile(cfgFile)
 			assert.NoError(err)
 			assert.Equal("staging", cfg.DefaultEnv)
 		})
@@ -620,7 +620,7 @@ func TestLoadEnv(t *testing.T) {
 			err = os.Setenv(tt.envNew, "4,5,6")
 			assert.NoError(err)
 			defer os.Unsetenv(tt.envNew)
-			_, err = loadConfigFile("./testdata/full.yaml")
+			_, err = LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			if tt.envNew == "DD_APM_IGNORE_RESOURCES" {
 				assert.Equal([]string{"4", "5", "6"}, coreconfig.Datadog.GetStringSlice(tt.key))
@@ -637,7 +637,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "123")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("123", cfg.Endpoints[0].APIKey)
 	})
@@ -649,7 +649,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-site.com")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/undocumented.yaml")
+		cfg, err := LoadConfigFile("./testdata/undocumented.yaml")
 		assert.NoError(err)
 		assert.Equal(apiEndpointPrefix+"my-site.com", cfg.Endpoints[0].Host)
 	})
@@ -661,7 +661,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "true")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.True(cfg.Enabled)
 	})
@@ -673,7 +673,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-site.com")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-site.com", cfg.Endpoints[0].Host)
 	})
@@ -685,7 +685,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-proxy.url")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-proxy.url", cfg.ProxyURL.String())
 	})
@@ -697,7 +697,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-proxy.url")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-proxy.url", cfg.ProxyURL.String())
 	})
@@ -709,7 +709,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "local.host")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("local.host", cfg.Hostname)
 	})
@@ -721,7 +721,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "bindhost.com")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("bindhost.com", cfg.StatsdHost)
 	})
@@ -736,7 +736,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "1234")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(1234, cfg.ReceiverPort)
 		})
@@ -749,7 +749,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "4321")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal(4321, cfg.StatsdPort)
 	})
@@ -761,7 +761,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "true")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/undocumented.yaml")
+		cfg, err := LoadConfigFile("./testdata/undocumented.yaml")
 		assert.NoError(err)
 		assert.Equal("0.0.0.0", cfg.ReceiverHost)
 	})
@@ -776,7 +776,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "1,2,3")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal([]string{"1", "2", "3"}, cfg.Ignore["resource"])
 		})
@@ -789,7 +789,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "warn")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("warn", cfg.LogLevel)
 	})
@@ -801,7 +801,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "web|http.request=1,db|sql.query=0.5")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal(map[string]map[string]float64{
 			"web": {"http.request": 1},
@@ -816,7 +816,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, `[{"name":"name1", "pattern":"pattern1"}, {"name":"name2","pattern":"pattern2","repl":"replace2"}]`)
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		rule1 := &config.ReplaceRule{
 			Name:    "name1",
@@ -840,7 +840,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, `important1 important2:value1`)
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal(cfg.RequireTags, []*config.Tag{{K: "important1", V: ""}, {K: "important2", V: "value1"}})
 	})
@@ -852,7 +852,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, `bad1:value1`)
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal(cfg.RejectTags, []*config.Tag{{K: "bad1", V: "value1"}})
 	})
@@ -867,7 +867,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "50")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(50, cfg.ConnectionLimit)
 		})
@@ -883,7 +883,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "6")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(6., cfg.TargetTPS)
 		})
@@ -898,7 +898,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "12")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(12., cfg.ErrorTPS)
 		})
@@ -913,7 +913,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "true")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(true, cfg.DisableRareSampler)
 		})
@@ -929,7 +929,7 @@ func TestLoadEnv(t *testing.T) {
 			err := os.Setenv(envKey, "7")
 			assert.NoError(err)
 			defer os.Unsetenv(envKey)
-			cfg, err := loadConfigFile("./testdata/full.yaml")
+			cfg, err := LoadConfigFile("./testdata/full.yaml")
 			assert.NoError(err)
 			assert.Equal(7., cfg.MaxEPS)
 		})
@@ -942,7 +942,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "337.41")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal(337.41, cfg.MaxRemoteTPS)
 	})
@@ -954,7 +954,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, `{"url1": ["key1", "key2"], "url2": ["key3"]}`)
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		cfg, err := loadConfigFile("./testdata/full.yaml")
+		cfg, err := LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Contains(cfg.Endpoints, &config.Endpoint{APIKey: "key1", Host: "url1"})
 		assert.Contains(cfg.Endpoints, &config.Endpoint{APIKey: "key2", Host: "url1"})
@@ -968,7 +968,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-site.com")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-site.com", coreconfig.Datadog.GetString("apm_config.profiling_dd_url"))
 	})
@@ -980,7 +980,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-site.com")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-site.com", coreconfig.Datadog.GetString("apm_config.debugger_dd_url"))
 	})
@@ -992,7 +992,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "my-key")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.Equal("my-key", coreconfig.Datadog.GetString("apm_config.debugger_api_key"))
 	})
@@ -1004,7 +1004,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "false")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.False(coreconfig.Datadog.GetBool("apm_config.obfuscation.credit_cards.enabled"))
 	})
@@ -1016,7 +1016,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, "false")
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		assert.False(coreconfig.Datadog.GetBool("apm_config.obfuscation.credit_cards.luhn"))
 	})
@@ -1028,7 +1028,7 @@ func TestLoadEnv(t *testing.T) {
 		err := os.Setenv(env, `{"url1": ["key1", "key2"], "url2": ["key3"]}`)
 		assert.NoError(err)
 		defer os.Unsetenv(env)
-		_, err = loadConfigFile("./testdata/full.yaml")
+		_, err = LoadConfigFile("./testdata/full.yaml")
 		assert.NoError(err)
 		expected := map[string][]string{
 			"url1": {"key1", "key2"},
