+++ cmd/trace-agent/config/testdata/full.yaml
