+++ cmd/trace-agent/config/testdata/ok_env_apm_config.yaml
