+++ cmd/trace-agent/config/testdata/ok_env_host_tag.yaml
