+++ cmd/trace-agent/config/testdata/ok_env_top_level.yaml
