+++ cmd/trace-agent/config/testdata/stringcode.go.tmpl
