+++ cmd/trace-agent/config/testdata/undocumented.yaml
