+++ pkg/trace/api/version.go
@@ -74,7 +74,7 @@ const (
 	//
 	v05 Version = "v0.5"
 
-	// v07
+	// V07 API
 	//
 	// Content-Type: application/msgpack
 	// Payload: TracerPayload (pkg/trace/pb/tracer_payload.proto)
