+++ pkg/trace/config/config.go
@@ -89,6 +89,7 @@ type ObfuscationConfig struct {
 	CreditCards CreditCardsConfig `mapstructure:"credit_cards"`
 }
 
+// AppSecConfig ...
 type AppSecConfig struct {
 	// Enabled reports whether AppSec is enabled.
 	Enabled bool
@@ -381,25 +382,7 @@ type AgentConfig struct {
 	RemoteSamplingClient RemoteClient
 }
 
-type ProfilingSettings struct {
-	// ProfilingURL specifies the URL to which profiles will be sent.  This can be constructed
-	// from a site value with ProfilingURLTemplate.
-	ProfilingURL string
-	// Period specifies the interval at which to collect profiles.
-	Period time.Duration
-	// CPUDuration specifies the length at which to collect CPU profiles.
-	CPUDuration time.Duration
-	// MutexProfileFraction, if set, turns on mutex profiles with rate
-	// indicating the fraction of mutex contention events reported in the mutex
-	// profile.
-	MutexProfileFraction int
-	// BlockProfileRate turns on block profiles with the given rate.
-	BlockProfileRate int
-	// WithGoroutineProfile additionally reports stack traces of all current goroutines
-	WithGoroutineProfile bool
-}
-
-// Remote client is used to APM Sampling Updates from a remote source. Within the Datadog Agent
+// RemoteClient client is used to APM Sampling Updates from a remote source. Within the Datadog Agent
 // the implementation is (cmd/trace-agent.remoteClient).
 type RemoteClient interface {
 	SamplingUpdates() <-chan SamplingUpdate
